RavenDB-26798 - move majority write-assurance tests from ExternalReplicationTests to ReplicationWriteAssurance
